backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-book-content-processor"
version = "0.1.0"
description = "RAG system for processing book content and generating embeddings"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Claude Code User", email = "user@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "cohere>=4.0.0",
    "qdrant-client>=1.8.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.4.0"
]

[project.urls]
Homepage = "https://github.com/user/repo"
Repository = "https://github.com/user/repo"

[tool.setuptools.packages.find]
where = ["."]